Refactor validation to do more legible
