+++ src/validation/vessel.validation.js
@@ -1,72 +1,42 @@
-const Joi = require('@hapi/joi');
 const {
   errors: { UnprocessableEntityException },
 } = require('auth-middleware');
 const SqlWhereParser = require('sql-where-parser');
-const { VESSELS_CONSTANTS: {
-  DEFAULT_PROPERTY_SUGGEST, IMO, MMSI, SHIPNAME, VESSEL_ID, FLAG, CALLSIGN
+const { VESSELS_CONSTANTS: { IMO, MMSI, SHIPNAME, VESSEL_ID, FLAG, CALLSIGN } } = require('../constant');
+const { sanitizeSQL } = require('../utils/sanitize-sql');
+const { splitDatasets } = require('../utils/split-datasets');
+const {
+  schemaGetAllVesselsV1,
+  schemaGetVesselByIdV1,
+  schemaSearchVesselsV1
+} = require('./schemas/vessel.schemas');
+
+async function validateSchema(ctx, schema) {
+  const value = await schema.validateAsync(ctx.request.query);
+  Object.keys(value).forEach(k => {
+    ctx.query[k] = value[k];
+  });
 }
-} = require('../constant');
-const { sanitizeSQL } = require('../utils/sanitize-sql')
 
-function splitDatasets(datasetsFromQuery) {
-  return datasetsFromQuery
-    .split(',')
-    .map(d => (d.indexOf(':') === -1 ? `${d}:latest` : d));
-}
-
-const vesselDefault = {
-  offset: 0,
-  queryFields: [],
-  suggestField: DEFAULT_PROPERTY_SUGGEST,
-  querySuggestions: false,
-  limit: 10,
-  binary: false,
-};
-
-const schemaGetAllVesselsV1 = Joi.object({
-  limit: Joi.number()
-    .integer()
-    .min(1)
-    .max(25)
-    .default(vesselDefault.limit),
-    binary: Joi.boolean().default(vesselDefault.binary),
-  ids: Joi.string().required(),
-  offset: Joi.number()
-  .integer()
-  .min(0)
-  .default(vesselDefault.offset),
-  datasets: Joi.string().required(),
-});
 async function getAllVesselsV1Validation(ctx, next) {
   try {
-    const value = await schemaGetAllVesselsV1.validateAsync(ctx.request.query);
-    Object.keys(value).forEach(k => {
-      ctx.query[k] = value[k];
-    });
+    await validateSchema(ctx, schemaGetAllVesselsV1)
     if (ctx.query.ids && !Array.isArray(ctx.query.ids)) {
       ctx.query.ids = ctx.query.ids.split(',');
+    }
+    if (ctx.query.datasets) {
+      ctx.query.datasets = splitDatasets(ctx.query.datasets)
     }
   } catch (err) {
     throw new UnprocessableEntityException('Invalid query', err.details);
   }
-
-  if (ctx.query.datasets) {
-    ctx.query.datasets = splitDatasets(ctx.query.datasets)
-  }
   await next();
 }
 
-const schemaGetVesselByIdV1 = Joi.object({
-  binary: Joi.boolean().default(false),
-  datasets: Joi.string().required(),
-});
+
 async function getVesselByIdV1Validation(ctx, next) {
   try {
-    const value = await schemaGetVesselByIdV1.validateAsync(ctx.request.query);
-    Object.keys(value).forEach(k => {
-      ctx.query[k] = value[k];
-    });
+    await validateSchema(ctx, schemaGetVesselByIdV1)
   } catch (err) {
     throw new UnprocessableEntityException('Invalid query', err.details);
   }
@@ -84,91 +54,66 @@ async function getVesselByIdV1Validation(ctx, next) {
   await next();
 }
 
-const schemaSearchVesselsV1 = Joi.object({
-  limit: Joi.number()
-    .integer()
-    .min(1)
-    .max(25)
-    .default(vesselDefault.limit),
-  query: Joi.string().required(),
-  binary: Joi.boolean().default(vesselDefault.binary),
-  suggestField: Joi.string().default(vesselDefault.suggestField),
-  queryFields: Joi.string().default(vesselDefault.queryFields),
-  querySuggestions: Joi.boolean().default(vesselDefault.querySuggestions),
-  offset: Joi.number()
-    .integer()
-    .min(0)
-    .default(vesselDefault.offset),
-  datasets: Joi.string().required(),
-});
+
 async function searchVesselsV1Validation(ctx, next) {
   try {
-    const value = await schemaSearchVesselsV1.validateAsync(ctx.request.query);
-
-    Object.keys(value).forEach(k => {
-      ctx.query[k] = value[k];
-    });
+    await validateSchema(ctx, schemaSearchVesselsV1)
     if (ctx.query.queryFields && !Array.isArray(ctx.query.queryFields)) {
       ctx.query.queryFields = ctx.query.queryFields.split(',');
+    }
+    if (ctx.query.datasets) {
+      ctx.query.datasets = splitDatasets(ctx.query.datasets);
     }
   } catch (err) {
     throw new UnprocessableEntityException('Invalid query', err.details);
   }
-
-  if (ctx.query.datasets) {
-    ctx.query.datasets = ctx.query.datasets
-      .split(',')
-      .map(d => (d.indexOf(':') === -1 ? `${d}:latest` : d));
-  }
-
   await next();
 }
 
 // Advance search SQL validation
 function validateFields(where, fields) {
+
+  if (!where || typeof where === 'string') {
+    throw new UnprocessableEntityException('Invalid Query: ', {
+      message: 'Query malformed, remember to exclude where and use correct syntax.',
+      path: ['query'],
+    })
+  }
+
   if (Array.isArray(where)) {
     where.forEach((condition) => validateFields(condition, fields));
-  } else {
-    if (!where || typeof where === 'string') {
+    return;
+  }
+
+  Object.keys(where).forEach((k) => {
+    if (k.toLowerCase() === 'and' || k.toLowerCase() === 'or') {
+      validateFields(where[k], fields);
+      return;
+    }
+
+    if (Array.isArray(where[k]) && !where[k].some((c) => fields.indexOf(c) >= 0)) {
       throw new UnprocessableEntityException('Invalid Query: ', {
-        message: 'Query malformed, remember to exclude where and use correct syntax.',
+        message: `The column "${where[k][0].toUpperCase()}" is not supported to search. Supported columns: "${fields.map(f => f.toUpperCase())}"`,
         path: ['query'],
       })
     }
-    Object.keys(where).forEach((k) => {
-      if (k.toLowerCase() === 'and' || k.toLowerCase() === 'or') {
-        validateFields(where[k], fields);
-      } else if (Array.isArray(where[k])) {
-        if (
-          !where[k].some((c) => fields.indexOf(c) >= 0)
-        ) {
-          throw new UnprocessableEntityException('Invalid Query: ', {
-            message: `The column "${where[k][0].toUpperCase()}" is not supported to search. Supported columns: "${fields.map(f => f.toUpperCase())}"`,
-            path: ['query'],
-          })
-        }
-      } else {
-        validateFields(where[k], fields);
-      }
-    });
-  }
+    validateFields(where[k], fields);
+  });
+
 }
 async function advanceSearchSqlValidation(ctx, next) {
-  let { query: { query: sql } } = ctx;
-  sql = sanitizeSQL(sql)
-  const fieldsAllowed = [IMO, MMSI, SHIPNAME, VESSEL_ID, FLAG, CALLSIGN];
-  let whereParsed;
   try {
+    const { query: { query: sql } } = ctx;
     const parser = new SqlWhereParser();
-    whereParsed = parser.parse(sql);
+    const whereParsed = parser.parse(sanitizeSQL(sql));
+    validateFields(whereParsed, [IMO, MMSI, SHIPNAME, VESSEL_ID, FLAG, CALLSIGN]);
+    ctx.query.query.sql = sql;
   } catch (err) {
     throw new UnprocessableEntityException('Invalid Query: ', {
       message: `Query malformed, remember to exclude where and use correct syntax.`,
       path: ['query'],
     })
   }
-  validateFields(whereParsed, fieldsAllowed);
-  ctx.query.query.sql = sql;
   await next()
 }
